kimi-k3 ar_fusion: review cleanups (comment placement, eps type, docs)
- Move the SGLANG_K3_AR_FUSION o_proj comment in KimiK3DeltaAttention onto
  the reduce_results kwarg it describes (was stranded above conv_weights).
- pull_norm host entry: double eps -> float eps, matching push_norm and the
  deepseek_v4 wrapper convention.
- Fix the test docstring usage path.
- black: reformat two pre-existing long lines to match upstream kimi-k3.

+++ python/sglang/jit_kernel/csrc/kimi_k3/comm/ar_fusion.cuh
@@ -576,7 +576,7 @@ struct AllReduceFusionKernel {
         .enable_pdl(kUsePDL)(all_reduce_push_norm_cluster_kernel<kWorldSize, kClusterSize, kUsePDL>, params);
   }
 
-  static void pull_norm(CommunicatorRef ref, TensorView input, TensorView weight, double eps, int64_t input_mc_ptr) {
+  static void pull_norm(CommunicatorRef ref, TensorView input, TensorView weight, float eps, int64_t input_mc_ptr) {
     constexpr uint32_t kDefaultBlocks = 64;
     const auto& data = *ref.get();
     auto params = make_params_norm(data, input, weight, eps);
